improved regplot and bar_chart
