+++ easy_mpl/_main.py
@@ -1,15 +1,11 @@
 
 __all__ = [
     "plot",
-    "bar_chart",
-    "hist",
     "scatter",
-    "contour",
     "dumbbell_plot"
 ]
 
-import random
-from typing import Union, Tuple
+from typing import Tuple
 
 import numpy as np
 import pandas as pd
@@ -17,141 +13,7 @@ import matplotlib as mpl
 import matplotlib.pyplot as plt
 import matplotlib.lines as mlines
 
-from .utils import to_1d_array, make_cols_from_cmap, process_axis
-from .utils import BAR_CMAPS
-
-
-def bar_chart(
-        values,
-        labels=None,
-        orient='h',
-        sort=False,
-        errors: Union = None,
-        color=None,
-        cmap: str = None,
-        rotation=0,
-        show=True,
-        ax=None,
-        bar_labels: Union[list, np.ndarray] = None,
-        bar_label_kws=None,
-        ax_kws: dict = None,
-        **kwargs
-) -> plt.Axes:
-    """
-    plots bar chart
-
-    Parameters
-    -----------
-        values : array like
-            1 dimensional array or list
-        labels : list, optional
-            used for labeling each bar
-        orient : `str`, optional
-            orientation of bars. either 'h' or 'v'
-        sort : bool, optional
-            whether to sort the bars based upon their values or not
-        errors : list, optional
-            for error bars
-        color : bool, optional (default=None)
-            color for bars. It can any color value valid for matplotlib.
-        cmap : str, optional (default=None)
-            matplotlib colormap
-        rotation : int, optional
-            rotation angle of ticklabels
-        ax : plt.Axes, optional
-            If not given, current available axes will be used
-        show : bool, optional,
-        bar_labels : list
-            labels of the bars
-        bar_label_kws : dict
-        ax_kws : dict, optional
-            any keyword arguments for processing of axes that will go to
-            :py:func:`easy_mpl.utils.process_axes`
-        **kwargs :
-            any additional keyword arguments for `axes.bar`_ or `axes.barh`_
-
-    Returns
-    --------
-    matplotlib.pyplot.Axes
-        matplotlib Axes on which the bar_chart is drawn. If ``show`` is False, this axes
-        can be used for further processing
-
-    Example
-    --------
-        >>> from easy_mpl import bar_chart
-        >>> bar_chart([1,2,3,4,4,5,3,2,5])
-        specifying labels
-        >>> bar_chart([3,4,2,5,10], ['a', 'b', 'c', 'd', 'e'])
-        sorting the data
-        >>> bar_chart([1,2,3,4,4,5,3,2,5], sort=True)
-
-
-    .. _axes.bar:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.bar.html
-
-    .. _axes.barh:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.barh.html
-    """
-
-    if labels is None:
-        if isinstance(values, (pd.DataFrame, pd.Series)):
-            labels = values.index
-
-    values = to_1d_array(values)
-
-    cmap = make_cols_from_cmap(cmap or random.choice(BAR_CMAPS), len(values), 0.2)
-    color = color if color is not None else cmap
-
-    if not ax:
-        ax = plt.gca()
-        if 'figsize' in kwargs:
-            figsize = kwargs.pop('figsize')
-            ax.figure.set_size_inches(figsize)
-
-    if labels is None:
-        labels = [f"F{i}" for i in range(len(values))]
-
-    if sort:
-        sort_idx = np.argsort(values)
-        values = values[sort_idx]
-        labels = np.array(labels)[sort_idx]
-
-    if orient in ['h', 'horizontal']:
-        bar = ax.barh(np.arange(len(values)), values, color=color, **kwargs)
-        ax.set_yticks(np.arange(len(values)))
-        ax.set_yticklabels(labels, rotation=rotation)
-
-        if bar_labels is not None:
-            bar_label_kws = bar_label_kws or {'label_type': 'center'}
-            ax.bar_label(bar, labels=bar_labels, **bar_label_kws)
-
-        if errors is not None:
-            ax.errorbar(values, np.arange(len(values)), xerr=errors, fmt=".",
-                        color="black")
-
-    else:
-        bar = ax.bar(np.arange(len(values)), values, color=color, **kwargs)
-        ax.set_xticks(np.arange(len(values)))
-        ax.set_xticklabels(labels, rotation=rotation)
-
-        if bar_labels is not None:
-            bar_label_kws = bar_label_kws or {'label_type': 'center'}
-            ax.bar_label(bar, labels=bar_labels, **bar_label_kws)
-
-        if errors is not None:
-            ax.errorbar(np.arange(len(values)), values, yerr=errors, fmt=".",
-                        color="black")
-
-    if ax_kws:
-        ax = process_axis(ax, **ax_kws)
-
-    if 'label' in kwargs:
-        ax.legend()
-
-    if show:
-        plt.show()
-
-    return ax
+from .utils import to_1d_array, process_axis
 
 
 def plot(
@@ -289,63 +151,6 @@ def _process_axis(ax, show, kwargs):
     return ax
 
 
-def hist(
-        x: Union[list, np.ndarray, pd.Series, pd.DataFrame],
-        hist_kws: dict = None,
-        grid: bool = True,
-        ax: plt.Axes = None,
-        show: bool = True,
-        **kwargs
-) -> plt.Axes:
-    """
-    one stop shop for histogram
-
-    Parameters
-    -----------
-        x : list, array, optional
-            array like, must not be greader than 1d
-        grid : bool, optional
-            whether to show the grid or not
-        show : bool, optional
-            whether to show the plot or not
-        ax : plt.Axes, optional
-            axes on which to draw the plot
-        hist_kws : dict, optional
-            any keyword arguments for `axes.hist`_
-        **kwargs : optional
-            any keyword arguments for axes manipulation such as title, xlable, ylable etc
-
-    matplotlib.pyplot.Axes
-        matplotlib Axes on which the histogram is drawn. If ``show`` is False, this axes
-        can be used for further processing
-
-    Example
-    --------
-        >>> from easy_mpl import hist
-        >>> import numpy as np
-        >>> hist(np.random.random((10, 1)))
-
-    .. _axes.hist:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.hist.html
-    """
-
-    if not ax:
-        ax = plt.gca()
-        if 'figsize' in kwargs:
-            figsize = kwargs.pop('figsize')
-            ax.figure.set_size_inches(figsize)
-
-    hist_kws = hist_kws or {}
-    n, bins, patches = ax.hist(x, **hist_kws)
-
-    process_axis(ax, grid=grid, **kwargs)
-
-    if show:
-        plt.show()
-
-    return ax
-
-
 def scatter(
         x,
         y,
@@ -417,128 +222,6 @@ def scatter(
         plt.show()
 
     return ax, sc
-
-
-def contour(
-        x,
-        y,
-        z,
-        fill_between=False,
-        show_points=False,
-        colorbar=True,
-        label_contours=False,
-        contour_kws: dict = None,
-        fill_between_kws: dict = None,
-        show_points_kws: dict = None,
-        label_contour_kws: dict = None,
-        ax=None,
-        show=True,
-        **kwargs
-):
-    """A contour plot of irregularly spaced data coordinates.
-
-    Parameters
-    ---------
-        x : array, list
-            a 1d array defining grid positions along x-axis
-        y : array, list
-            a 1d array defining grid positions along y-axis
-        z : array, list
-            values on grids/points defined by x and y
-        fill_between : bool, optional
-            whether to fill the space between colors or not
-        show_points : bool, optional
-            whether to show the
-        label_contours : bool, optional
-            whether to label the contour lines or not
-        colorbar : bool, optional
-            whether to show the colorbar or not. Only valid if `fill_between` is
-            True
-        contour_kws : dict, optional
-            any keword argument for `axes.tricontour`_
-        fill_between_kws : dict, optional
-            any keword argument for `axes.tricontourf`_
-        show_points_kws : dict, optional
-            any keword argument for `axes.plot`_
-        label_contour_kws : dict, optional
-            any keword argument for `axes.clabel`_
-        ax : plt.Axes, optional
-            matplotlib axes to work on. If not given, current active axes will
-            be used.
-        show : bool, optional
-            whether to show the plot or not
-        **kwargs : optional
-            any keyword arguments for easy_mpl.utils.process_axis
-
-    Returns
-    -------
-        a matplotliblib Axes
-
-    Examples
-    --------
-        >>> from easy_mpl import contour
-        >>> import numpy as np
-        >>> _x = np.random.uniform(-2, 2, 200)
-        >>> _y = np.random.uniform(-2, 2, 200)
-        >>> _z = _x * np.exp(-_x**2 - _y**2)
-        >>> contour(_x, _y, _z, fill_between=True, show_points=True)
-        ... # show contour labels
-        >>> contour(_x, _y, _z, label_contours=True, show_points=True)
-
-    Note
-    ----
-    The length of x and y should be same. The actual grid is created using axes.tricontour
-    and axes.tricontourf functions.
-
-    .. _axes.tricontour:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.tricontour.html
-
-    .. _axes.tricontourf:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.tricontourf.html
-
-    .. _axes.plot:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.plot.html
-
-    .. _axes.clabel:
-        https://matplotlib.org/stable/api/_as_gen/matplotlib.axes.Axes.clabel.html
-    """
-    x = np.array(x)
-    y = np.array(y)
-    z = np.array(z)
-
-    assert len(x) == len(y) == len(z)
-
-    if ax is None:
-        ax = plt.gca()
-        if 'figsize' in kwargs:
-            figsize = kwargs.pop('figsize')
-            ax.figure.set_size_inches(figsize)
-
-    kws = contour_kws or {"levels": 14, "linewidth": 0.5, "colors": "k"}
-    CS = ax.tricontour(x, y, z, **kws)
-
-    if fill_between:
-        kws = fill_between_kws or {'levels': 14, 'cmap': 'RdBu_r'}
-        CS = ax.tricontourf(x, y, z, **kws)
-
-        if colorbar:
-            fig: plt.Figure = plt.gcf()
-            fig.colorbar(CS, ax=ax)
-
-    if show_points:
-        kws = show_points_kws or {'color': 'k', 'marker': 'o', 'ms': 3, "linestyle": ""}
-        ax.plot(x, y, **kws)
-
-    process_axis(ax, **kwargs)
-
-    if label_contours:
-        kws = label_contour_kws or {"colors": "black"}
-        ax.clabel(CS, **kws)
-
-    if show:
-        plt.show()
-
-    return ax
 
 
 def dumbbell_plot(
