+++ easy_mpl/utils.py
@@ -26,78 +26,6 @@ RIDGE_CMAPS = [
     "RdBu", "Spectral",
 ]
 
-regplot_combs = [
-    ['cadetblue', 'slateblue', 'darkslateblue'],
-    ['cadetblue', 'mediumblue', 'mediumblue'],
-    ['cornflowerblue', 'dodgerblue', 'darkblue'],
-    ['cornflowerblue', 'dodgerblue', 'steelblue'],
-    ['cornflowerblue', 'mediumblue', 'dodgerblue'],
-    ['cornflowerblue', 'steelblue', 'mediumblue'],
-    ['darkslateblue', 'aliceblue', 'mediumblue'],
-    ['darkslateblue', 'blue', 'royalblue'],
-    ['darkslateblue', 'blueviolet', 'royalblue'],
-    ['darkslateblue', 'darkblue', 'midnightblue'],
-    ['darkslateblue', 'mediumblue', 'darkslateblue'],
-    ['darkslateblue', 'midnightblue', 'mediumblue'],
-    ['seagreen', 'darkslateblue', 'cadetblue'],
-    ['cadetblue', 'darkblue', 'midnightblue'],
-    ['cadetblue', 'deepskyblue', 'cadetblue']
-]
-
-
-def _regplot_paras(x, y, ci:int=None):
-    """prepares parameters for regplot"""
-    grid = np.linspace(np.min(x), np.max(x), 100)
-    x = np.c_[np.ones(len(x)), x]
-    grid = np.c_[np.ones(len(grid)), grid]
-    yhat = grid.dot(reg_func(x, y))
-
-    err_bands = None
-    if ci:
-        boots = bootdist(reg_func, args=[x, y], n_boot=1000).T
-
-        yhat_boots = grid.dot(boots).T
-        err_bands = _ci(yhat_boots, ci, axis=0)
-
-    return grid, yhat, err_bands
-
-
-def _regplot(x, y, ax, ci=None, line_color=None, fill_color=None):
-
-    grid, yhat, err_bands = _regplot_paras(x, y, ci)
-
-    ax.plot(grid[:, 1], yhat, color=line_color)
-
-    if ci:
-        ax.fill_between(grid[:, 1], *err_bands,
-                        facecolor=fill_color,
-                        alpha=.15)
-    return ax
-
-
-def _ci(a, which=95, axis=None):
-    """Return a percentile range from an array of values."""
-    p = 50 - which / 2, 50 + which / 2
-    return np.nanpercentile(a, p, axis)
-
-
-def reg_func(_x, _y):
-    return np.linalg.pinv(_x).dot(_y)
-
-
-def bootdist(f, args, n_boot=1000, **func_kwargs):
-
-    n = len(args[0])
-    integers = np.random.randint
-    boot_dist = []
-    for i in range(int(n_boot)):
-        resampler = integers(0, n, n, dtype=np.intp)  # intp is indexing dtype
-        sample = [a.take(resampler, axis=0) for a in args]
-        boot_dist.append(f(*sample, **func_kwargs))
-
-    return np.array(boot_dist)
-
-
 def process_axis(
         ax: plt.Axes=None,
         label:str = None,
